requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
duckduckgo_search==8.1.1
faiss-cpu==1.11.0.post1
GitPython==3.1.45
httpx==0.28.1
httpx-sse==0.4.1
Jinja2==3.1.6
jsonschema==4.25.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langgraph==0.6.4
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.13
lxml==6.0.0
marshmallow==3.26.1
numpy==2.3.2
openai==1.99.8
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
primp==0.15.0
propcache==0.3.2
protobuf==6.31.1
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydeck==0.9.1
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
SQLAlchemy==2.0.43
sqlite-vec==0.1.6
streamlit==1.48.0
tenacity==9.1.2
tiktoken==0.11.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.14.1
xxhash==3.5.0
yarl==1.20.1
youtube-transcript-api==0.6.2
zstandard==0.23.0
